+++ src/edu/ujcv/progra1/SortTester.java
@@ -1,5 +1,5 @@
 package edu.ujcv.progra1;
 
 public interface SortTester {
-    public long sort(int[] array);
+    public double sort(int[] array);
 }
